# Error when negative weights or zero sum are used when sampling