@bkamins 2022-09-04
@nalimilan - do we have benchmarks showing the performance impact?

Also - given the changes we do I think it would be good in https://juliastats.org/StatsBase.jl/stable/weights/#Weight-Vectors-1 to add a recommendation to create weights vector once if it is reused later. The point is that the code like:
```
[sample(Weights(x)) for i in 1:10^6]
```
is inefficient as it should be replaced by:
```
w = Weights(x)
[sample(w) for i in 1:10^6]
```
and I commonly see such patterns since users are unaware of the cost of `Weights` creation.