@nalimilan 2022-09-04
Good catch. This seems to be due to `direct_sample!` calling `sample(rng, wv)` for each requested value, which means we check aliasing, sum and negativity of weights 100 times. The code was maybe efficient when it was written, but we've progressively added sanity checks and now it's really absurd. Looks like by copying the code I can almost eliminate the overhead. I'll do more investigations.